+++ com.unity.render-pipelines.lightweight/Runtime/Data/LightweightRenderPipelineAsset.cs
@@ -447,6 +447,21 @@ namespace UnityEngine.Rendering.LWRP
         {
             get { return editorResources.autodeskInteractiveMaskedShader; }
         }
+
+        public override Shader terrainDetailLitShader
+        {
+            get { return editorResources.terrainDetailLitShader; }
+        }
+
+        public override Shader terrainDetailGrassShader
+        {
+            get { return editorResources.terrainDetailGrassShader; }
+        }
+
+        public override Shader terrainDetailGrassBillboardShader
+        {
+            get { return editorResources.terrainDetailGrassBillboardShader; }
+        }
 #endif
 
         public Shader blitShader